# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospital CRM & Patient Engagement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors
      parameters:
        - name: specialization
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Create a doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorInput"
      responses:
        "201":
          description: Created doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorUpdate"
      responses:
        "200":
          description: Updated doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    delete:
      operationId: deleteDoctor
      tags: [doctors]
      summary: Delete a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Contact Variables ──────────────────────────────────────────────────────
...
  /dashboard/home:
    get:
      operationId: getHomeDashboard
      tags: [dashboard]
      summary: Collective home dashboard data
      parameters:
        - name: channel
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Home dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeDashboard"
  /dashboard/crm:
    get:
      operationId: getCrmDashboard
      tags: [dashboard]
      summary: CRM dashboard data
      parameters:
        - name: channel
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CRM dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmDashboard"
...
  /dashboard/appointments:
    get:
      operationId: getAppointmentsDashboard
      tags: [dashboard]
      summary: Appointments dashboard data
      parameters:
        - name: channel
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Appointments dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsDashboard"
...
components:
  schemas:
...
    Doctor:
      type: object
      required: [id, name, specialization, active]
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
        qualifications:
          type: ["string", "null"]
        experience:
          type: ["integer", "null"]
        active:
          type: boolean
        slots:
          type: array
          items:
            $ref: "#/components/schemas/DoctorSlot"
    DoctorSlot:
      type: object
      required: [day, startTime, endTime]
      properties:
        day:
          type: string
        startTime:
          type: string
        endTime:
          type: string
    DoctorInput:
      type: object
      required: [name, specialization]
      properties:
        name:
          type: string
        specialization:
          type: string
        qualifications:
          type: string
        experience:
          type: integer
        slots:
          type: array
          items:
            $ref: "#/components/schemas/DoctorSlot"
    DoctorUpdate:
      type: object
      properties:
        name:
          type: string
        specialization:
          type: string
        qualifications:
          type: string
        experience:
          type: integer
        active:
          type: boolean
        slots:
          type: array
          items:
            $ref: "#/components/schemas/DoctorSlot"

    # ── Contact Variable ──────────────────────────────────────────────────────
...
    Lead:
      type: object
      required: [id, firstName, lastName, mobile, sourceChannel, status, optedIn, dndListed, createdAt, lastActionAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        mobile:
          type: string
        uhid:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        sourceChannel:
          type: string
          enum: [waba, web_chat, form, csv, app_booking, web_booking, push]
        status:
          type: string
          enum: [new, contacted, in_progress, fulfilled, closed]
        ownerUserId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        sourceListTag:
          type: ["string", "null"]
        optedIn:
          type: boolean
        dndListed:
          type: boolean
        createdAt:
          type: string
        lastActionAt:
          type: string
        hasActiveSession:
          type: boolean
...
    Appointment:
      type: object
      required: [id, leadId, doctorId, specialization, sourceChannel, status, datetime, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        leadName:
          type: ["string", "null"]
        leadMobile:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        specialization:
          type: string
        sourceChannel:
          type: string
          enum: [app_booking, web_booking]
        status:
          type: string
          enum: [booked, confirmed, completed, cancelled]
        datetime:
          type: string
        createdAt:
          type: string
...
    LeadsByChannel:
      type: object
      required: [channel, count]
      properties:
        channel:
          type: string
        count:
          type: integer
    LeadsByStatus:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
    HomeDashboard:
      type: object
      required: [totalLeads, activeLeads, fulfilledLeads, leadsByChannel, leadsByStatus, conversionByChannel, recentLeads]
      properties:
        totalLeads:
          type: integer
        activeLeads:
          type: integer
        fulfilledLeads:
          type: integer
        leadsByChannel:
          type: array
          items:
            $ref: "#/components/schemas/LeadsByChannel"
        leadsByStatus:
          type: array
          items:
            $ref: "#/components/schemas/LeadsByStatus"
        conversionByChannel:
          type: array
          items:
            type: object
            required: [channel, total, converted, rate]
            properties:
              channel:
                type: string
              total:
                type: integer
              converted:
                type: integer
              rate:
                type: number
        revenueAttributed:
          type: number
        recentLeads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        appointmentSummary:
          type: object
          properties:
            total:
              type: integer
            today:
              type: integer
            upcoming:
              type: integer
    CrmDashboard:
      type: object
      required: [totalLeads, leadsByChannel, volumeTrend, slaBreaches]
      properties:
        totalLeads:
          type: integer
        leadsByChannel:
          type: array
          items:
            $ref: "#/components/schemas/LeadsByChannel"
        volumeTrend:
          type: array
          items:
            type: object
            required: [date, count]
            properties:
              date:
                type: string
              count:
                type: integer
        conversionTrend:
          type: array
          items:
            type: object
            required: [date, converted]
            properties:
              date:
                type: string
              converted:
                type: integer
        slaBreaches:
          type: integer
        agedLeads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
...
    AppointmentsDashboard:
      type: object
      required: [total, byChannel, byStatus, bySpecialization, recentAppointments]
      properties:
        total:
          type: integer
        today:
          type: integer
        byChannel:
          type: array
          items:
            type: object
            required: [channel, count]
            properties:
              channel:
                type: string
              count:
                type: integer
        byStatus:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                type: string
              count:
                type: integer
        bySpecialization:
          type: array
          items:
            type: object
            required: [specialization, count]
            properties:
              specialization:
                type: string
              count:
                type: integer
        trendByDay:
          type: array
          items:
            type: object
            required: [date, count]
            properties:
              date:
                type: string
              count:
                type: integer
        recentAppointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
...